fix: rename based on feedback
