+++ go.work
@@ -4,5 +4,5 @@ use (
 	./node
 	./api
 	./delegator
-	./sentry
+	./sentinel
 )
