+++ sentinel/.gitignore
