+++ sentinel/pkg/checker/event/types.go
