+++ sentinel/pkg/checker/health/cypress_healthcheck.json
