+++ sentinel/script/collect-healthcheck-urls/main.go
